add sync module
