+++ src/esthri/src/ops/mod.rs
@@ -1,2 +1,3 @@
 pub(super) mod download;
 pub(super) mod upload;
+pub(super) mod sync;
